fmt: change int style from long to short

+++ src/implementations/CollarPool.sol
@@ -25,7 +25,7 @@ abstract contract BaseCollarPoolState {
 
     // represents one partitioned slice of the liquidity pool, its providers, and the amount they provide
     struct Slot {
-        uint256 liquidity; // <-- total liquidity in the slot
+        uint liquidity; // <-- total liquidity in the slot
         EnumerableMap.AddressToUintMap providers;
     }
 
@@ -51,13 +51,13 @@ abstract contract BaseCollarPoolState {
     //  - has a defined time horizon
     //  - carries inherent risk
     struct Position {
-        uint256 expiration; // <-- defined time horizon --- does not change
-        uint256 principal; // <-- static initial value --- does not change
-        uint256 withdrawable; // <-- zero until close, then set to settlement value
+        uint expiration; // <-- defined time horizon --- does not change
+        uint principal; // <-- static initial value --- does not change
+        uint withdrawable; // <-- zero until close, then set to settlement value
     }
 
     /// @notice Records the state of each slot (see Slot struct above)
-    mapping(uint256 index => Slot) internal slots;
+    mapping(uint index => Slot) internal slots;
 
     /// @notice Records the state of each Position by their UUID
     mapping(bytes32 uuid => Position) public positions;
@@ -75,12 +75,12 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
     // ----- IMMUTABLES ----- //
 
     /// @notice This is the ID of the slot that is unallocated to any particular call strike percentage
-    uint256 public constant UNALLOCATED_SLOT = type(uint256).max;
+    uint public constant UNALLOCATED_SLOT = type(uint).max;
 
     /// @notice This is the factor by which the slot ID is scaled to get the actual bps value
     /// @dev A tick scale factor of 1 means that slot 10_000 = 100% = 1 bps per tick, and so on
 
-    uint256 public immutable tickScaleFactor;
+    uint public immutable tickScaleFactor;
 
     /// @notice The address of the engine is set upon pool creation (and registered with the engine)
     address public immutable engine;
@@ -94,35 +94,35 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
     address public immutable collateralAsset;
 
     /// @notice The duration of collars to be opened against this pool
-    uint256 public immutable duration;
+    uint public immutable duration;
 
     /// @notice The LTV of collars to be opened against this pool
-    uint256 public immutable ltv;
+    uint public immutable ltv;
 
     // ----- STATE VARIABLES ----- //
 
     /// @notice The total amount of liquidity in the pool
     /// @dev Total liquidity = locked + free + redeemable
-    uint256 public totalLiquidity;
+    uint public totalLiquidity;
 
     /// @notice The amount of locked liquidity in the pool
-    uint256 public lockedLiquidity;
+    uint public lockedLiquidity;
 
     /// @notice The amount of free liquidity in the pool
-    uint256 public freeLiquidity;
+    uint public freeLiquidity;
 
     /// @notice The amount of redeemable liquidity in the pool
-    uint256 public redeemableLiquidity;
+    uint public redeemableLiquidity;
 
     // ----- CONSTRUCTOR ----- //
 
     constructor(
         address _engine,
-        uint256 _tickScaleFactor,
+        uint _tickScaleFactor,
         address _cashAsset,
         address _collateralAsset,
-        uint256 _duration,
-        uint256 _ltv
+        uint _duration,
+        uint _ltv
     ) {
         if (!CollarEngine(_engine).isValidLTV(_ltv)) {
             revert InvalidLTV();
@@ -138,68 +138,60 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
 
     // ----- VIEW FUNCTIONS ----- //
 
-    function getInitializedSlotIndices() external view override returns (uint256[] memory) {
+    function getInitializedSlotIndices() external view override returns (uint[] memory) {
         return initializedSlotIndices.values();
     }
 
-    function getLiquidityForSlots(uint256[] calldata slotIndices)
+    function getLiquidityForSlots(uint[] calldata slotIndices)
         external
         view
         override
-        returns (uint256[] memory)
+        returns (uint[] memory)
     {
-        uint256[] memory liquidity = new uint256[](slotIndices.length);
+        uint[] memory liquidity = new uint[](slotIndices.length);
 
-        for (uint256 i = 0; i < slotIndices.length; i++) {
+        for (uint i = 0; i < slotIndices.length; i++) {
             liquidity[i] = slots[slotIndices[i]].liquidity;
         }
 
         return liquidity;
     }
 
-    function getLiquidityForSlot(uint256 slotIndex) external view override returns (uint256) {
+    function getLiquidityForSlot(uint slotIndex) external view override returns (uint) {
         return slots[slotIndex].liquidity;
     }
 
-    function getNumProvidersInSlot(uint256 slotIndex) external view override returns (uint256) {
+    function getNumProvidersInSlot(uint slotIndex) external view override returns (uint) {
         return slots[slotIndex].providers.length();
     }
 
     function getSlotProviderInfoAtIndex(
-        uint256 slotIndex,
-        uint256 providerIndex
+        uint slotIndex,
+        uint providerIndex
     )
         external
         view
         override
-        returns (address, uint256)
+        returns (address, uint)
     {
         return slots[slotIndex].providers.at(providerIndex);
     }
 
     function getSlotProviderInfoForAddress(
-        uint256 slotIndex,
+        uint slotIndex,
         address provider
     )
         external
         view
         override
-        returns (uint256)
+        returns (uint)
     {
         return slots[slotIndex].providers.get(provider);
     }
 
-    function previewRedeem(
-        bytes32 uuid,
-        uint256 amount
-    )
-        public
-        view
-        override
-        returns (uint256 cashReceived)
-    {
+    function previewRedeem(bytes32 uuid, uint amount) public view override returns (uint cashReceived) {
         // verify that the user has enough tokens for this to even work
-        if (ERC6909TokenSupply(address(this)).balanceOf(msg.sender, uint256(uuid)) < amount) {
+        if (ERC6909TokenSupply(address(this)).balanceOf(msg.sender, uint(uuid)) < amount) {
             revert InvalidAmount();
         }
 
@@ -210,8 +202,8 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
             // if finalized, calculate final redeem value
             // grab collateral asset value @ exact vault expiration time
 
-            uint256 _totalTokenCashSupply = _position.withdrawable;
-            uint256 _totalTokenSupply = totalSupply[uint256(uuid)];
+            uint _totalTokenCashSupply = _position.withdrawable;
+            uint _totalTokenSupply = totalSupply[uint(uuid)];
 
             cashReceived = (_totalTokenCashSupply * amount) / _totalTokenSupply;
         } else {
@@ -228,7 +220,7 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
 
     // ----- STATE CHANGING FUNCTIONS ----- //
 
-    function addLiquidityToSlot(uint256 slotIndex, uint256 amount) public virtual override {
+    function addLiquidityToSlot(uint slotIndex, uint amount) public virtual override {
         Slot storage slot = slots[slotIndex];
 
         // If this slot isn't initialized, add to the initialized list - we're initializing it now
@@ -240,7 +232,7 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
             _allocate(slotIndex, msg.sender, amount);
         } else {
             address smallestProvider = _getSmallestProvider(slotIndex);
-            uint256 smallestAmount = slot.providers.get(smallestProvider);
+            uint smallestAmount = slot.providers.get(smallestProvider);
 
             if (smallestAmount > amount) revert NoLiquiditySpace();
 
@@ -259,11 +251,11 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         IERC20(cashAsset).transferFrom(msg.sender, address(this), amount);
     }
 
-    function withdrawLiquidityFromSlot(uint256 slotIndex, uint256 amount) public virtual override {
+    function withdrawLiquidityFromSlot(uint slotIndex, uint amount) public virtual override {
         console.log("withdraw liquidity from slot %d , amount  %d", slotIndex, amount);
         Slot storage slot = slots[slotIndex];
 
-        uint256 liquidity = slot.providers.get(msg.sender);
+        uint liquidity = slot.providers.get(msg.sender);
 
         // verify sender has enough liquidity in slot
         if (liquidity < amount) {
@@ -287,9 +279,9 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
     }
 
     function moveLiquidityFromSlot(
-        uint256 sourceSlotIndex,
-        uint256 destinationSlotIndex,
-        uint256 amount
+        uint sourceSlotIndex,
+        uint destinationSlotIndex,
+        uint amount
     )
         external
         virtual
@@ -303,7 +295,7 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         // withdrawLiquidityFromSlot(sourceSlotIndex, amount);
         // verify sender has enough liquidity in slot
         Slot storage sourceSlot = slots[sourceSlotIndex];
-        uint256 liquidity = sourceSlot.providers.get(msg.sender);
+        uint liquidity = sourceSlot.providers.get(msg.sender);
         if (liquidity < amount) {
             revert InvalidAmount();
         }
@@ -323,7 +315,7 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
             _allocate(destinationSlotIndex, msg.sender, amount);
         } else {
             address smallestProvider = _getSmallestProvider(destinationSlotIndex);
-            uint256 smallestAmount = destinationSlot.providers.get(smallestProvider);
+            uint smallestAmount = destinationSlot.providers.get(smallestProvider);
 
             if (smallestAmount > amount) revert NoLiquiditySpace();
 
@@ -333,15 +325,7 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         emit LiquidityMoved(msg.sender, sourceSlotIndex, destinationSlotIndex, amount);
     }
 
-    function openPosition(
-        bytes32 uuid,
-        uint256 slotIndex,
-        uint256 amount,
-        uint256 expiration
-    )
-        external
-        override
-    {
+    function openPosition(bytes32 uuid, uint slotIndex, uint amount, uint expiration) external override {
         // ensure this is a valid vault calling us - it must call through the engine
         if (!CollarEngine(engine).isVaultManager(msg.sender)) {
             revert NotCollarVaultManager();
@@ -349,7 +333,7 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
 
         // grab the slot
         Slot storage slot = slots[slotIndex];
-        uint256 numProviders = slot.providers.length();
+        uint numProviders = slot.providers.length();
 
         // if no providers, revert
         if (numProviders == 0) {
@@ -361,23 +345,23 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
             revert InvalidAmount();
         }
 
-        for (uint256 i = 0; i < numProviders; i++) {
+        for (uint i = 0; i < numProviders; i++) {
             // calculate how much to pull from provider based off of their proportional ownership of liquidity
             // in this slot
 
-            (address thisProvider, uint256 thisLiquidity) = slot.providers.at(i);
+            (address thisProvider, uint thisLiquidity) = slot.providers.at(i);
 
             // this provider's liquidity to pull =
             // (provider's proportional ownership of slot liquidity) * (total amount needed)
             // (providerLiquidity / totalSlotLiquidity) * amount
             // (providerLiquidity * amount) / totalSlotLiquidity
-            uint256 amountFromThisProvider = (thisLiquidity * amount) / slot.liquidity;
+            uint amountFromThisProvider = (thisLiquidity * amount) / slot.liquidity;
 
             // decrement the amount of free liquidity that this provider has, in this slot
             slot.providers.set(thisProvider, thisLiquidity - amountFromThisProvider);
 
             // mint tokens representing the provider's share in this vault to this provider
-            _mint(thisProvider, uint256(uuid), amountFromThisProvider);
+            _mint(thisProvider, uint(uuid), amountFromThisProvider);
 
             emit PoolTokensIssued(thisProvider, expiration, amountFromThisProvider);
         }
@@ -408,30 +392,30 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         emit PositionOpened(msg.sender, uuid, expiration, amount);
     }
 
-    function finalizePosition(bytes32 uuid, address vaultManager, int256 positionNet) external override {
+    function finalizePosition(bytes32 uuid, address vaultManager, int positionNet) external override {
         // verify caller via engine
         if (!CollarEngine(engine).isVaultManager(msg.sender)) {
             revert NotCollarVaultManager();
         }
 
         // either case, we need to set the withdrawable amount to principle + positionNet
-        positions[uuid].withdrawable = uint256(int256(positions[uuid].principal) + positionNet);
+        positions[uuid].withdrawable = uint(int(positions[uuid].principal) + positionNet);
 
         // update global liquidity amounts
         // free liquidity unchanged
 
-        totalLiquidity = uint256(int256(totalLiquidity) + positionNet);
+        totalLiquidity = uint(int(totalLiquidity) + positionNet);
 
         lockedLiquidity -= positions[uuid].principal;
 
-        redeemableLiquidity += uint256(int256(positions[uuid].principal) + positionNet);
+        redeemableLiquidity += uint(int(positions[uuid].principal) + positionNet);
 
         if (positionNet < 0) {
             // we owe the vault some tokens
-            IERC20(cashAsset).transfer(vaultManager, uint256(-positionNet));
+            IERC20(cashAsset).transfer(vaultManager, uint(-positionNet));
         } else if (positionNet > 0) {
             // the vault owes us some tokens
-            IERC20(cashAsset).transferFrom(vaultManager, address(this), uint256(positionNet));
+            IERC20(cashAsset).transferFrom(vaultManager, address(this), uint(positionNet));
         } else {
             // impressive. most impressive.
         }
@@ -439,7 +423,7 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         emit PositionFinalized(vaultManager, uuid, positionNet);
     }
 
-    function redeem(bytes32 uuid, uint256 amount) external override {
+    function redeem(bytes32 uuid, uint amount) external override {
         // validate position exists
         if (positions[uuid].expiration == 0) {
             revert InvalidVault();
@@ -450,12 +434,12 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         }
 
         // ensure that the user has enough tokens
-        if (ERC6909TokenSupply(address(this)).balanceOf(msg.sender, uint256(uuid)) < amount) {
+        if (ERC6909TokenSupply(address(this)).balanceOf(msg.sender, uint(uuid)) < amount) {
             revert InvalidAmount();
         }
 
         // calculate cash redeem value
-        uint256 redeemValue = previewRedeem(uuid, amount);
+        uint redeemValue = previewRedeem(uuid, amount);
 
         // adjust total redeemable cash
         positions[uuid].withdrawable -= redeemValue;
@@ -470,17 +454,17 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         emit Redemption(msg.sender, uuid, amount, redeemValue);
 
         // redeem to user & burn tokens
-        _burn(msg.sender, uint256(uuid), amount);
+        _burn(msg.sender, uint(uuid), amount);
         IERC20(cashAsset).transfer(msg.sender, redeemValue);
     }
 
     // ----- INTERNAL FUNCTIONS ----- //
 
-    function _isSlotInitialized(uint256 slotID) internal view returns (bool) {
+    function _isSlotInitialized(uint slotID) internal view returns (bool) {
         return initializedSlotIndices.contains(slotID);
     }
 
-    function _isSlotFull(uint256 slotID) internal view returns (bool full) {
+    function _isSlotFull(uint slotID) internal view returns (bool full) {
         if (slots[slotID].providers.length() == 5) {
             return true;
         } else {
@@ -488,15 +472,15 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         }
     }
 
-    function _getSmallestProvider(uint256 slotID) internal view returns (address smallestProvider) {
+    function _getSmallestProvider(uint slotID) internal view returns (address smallestProvider) {
         if (!_isSlotFull(slotID)) {
             return address(0);
         } else {
             Slot storage slot = slots[slotID];
-            uint256 smallestAmount = type(uint256).max;
+            uint smallestAmount = type(uint).max;
 
-            for (uint256 i = 0; i < 5; i++) {
-                (address _provider, uint256 _amount) = slot.providers.at(i);
+            for (uint i = 0; i < 5; i++) {
+                (address _provider, uint _amount) = slot.providers.at(i);
 
                 if (_amount < smallestAmount) {
                     smallestAmount = _amount;
@@ -506,11 +490,11 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         }
     }
 
-    function _allocate(uint256 slotID, address provider, uint256 amount) internal {
+    function _allocate(uint slotID, address provider, uint amount) internal {
         Slot storage slot = slots[slotID];
 
         if (slot.providers.contains(provider)) {
-            uint256 providerAmount = slot.providers.get(provider);
+            uint providerAmount = slot.providers.get(provider);
             slot.providers.set(provider, providerAmount + amount);
         } else {
             slot.providers.set(provider, amount);
@@ -519,10 +503,10 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         slot.liquidity += amount;
     }
 
-    function _unallocate(uint256 slotID, address provider, uint256 amount) internal {
+    function _unallocate(uint slotID, address provider, uint amount) internal {
         Slot storage slot = slots[slotID];
 
-        uint256 sourceAmount = slot.providers.get(provider);
+        uint sourceAmount = slot.providers.get(provider);
 
         if (sourceAmount == amount) slot.providers.remove(provider);
         else slot.providers.set(provider, sourceAmount - amount);
@@ -530,24 +514,17 @@ contract CollarPool is BaseCollarPoolState, ERC6909TokenSupply, ICollarPool {
         slot.liquidity -= amount;
     }
 
-    function _reAllocate(
-        address provider,
-        uint256 sourceSlotID,
-        uint256 destinationSlotID,
-        uint256 amount
-    )
-        internal
-    {
+    function _reAllocate(address provider, uint sourceSlotID, uint destinationSlotID, uint amount) internal {
         _unallocate(sourceSlotID, provider, amount);
         _allocate(destinationSlotID, provider, amount);
     }
 
-    function _mint(address account, uint256 id, uint256 amount) internal {
+    function _mint(address account, uint id, uint amount) internal {
         balanceOf[account][id] += amount;
         totalSupply[id] += amount;
     }
 
-    function _burn(address account, uint256 id, uint256 amount) internal {
+    function _burn(address account, uint id, uint amount) internal {
         balanceOf[account][id] -= amount;
         totalSupply[id] -= amount;
     }
